% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculate Score}
\usage{
calculate_score(
  site_id,
  parameter,
  unit,
  depth = NA,
  score_max,
  score_min,
  score_mean,
  score_median
)
}
\arguments{
\item{site_id}{Site ID.}

\item{parameter}{Parameter.}

\item{unit}{Parameter unit.}

\item{depth}{Depth category. Default value NA.}

\item{score_max}{Maximum score.}

\item{score_min}{Minimum score.}

\item{score_mean}{Average score.}

\item{score_median}{Median score.}
}
\value{
List including the annual value (score_num), way in which the annual
value was calculated (score_typ), and a category_score (score_str).
}
\description{
`calculate_score()` assigns numeric and categorical scores for
each parameter. Values are calculated by comparing the input data to the
threshold values for the relevant site, sampling depth, and parameter.

The numeric score returns the annual minimum, maximum, median, or mean value.
If the threshold table does not specify which value to return, the numeric
score uses the mean value.

The categorical score compares the numeric score to the threshold table and
returns a value of Excellent, Good, Fair, Poor, Does Not Meet Criteria, or
Meets Criteria.
}
