% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_thresholds}
\alias{dat_thresholds}
\title{Thresholds (dataframe)}
\format{
A dataframe with 13 columns:
 \describe{
   \item{State}{State. Values: CT, MA, RI, NA}
   \item{Group}{Group. Values: Coldwater, Saltwater, NA}
   \item{Site}{Site ID}
   \item{Depth}{Sample depth. Values: Surface, NA}
   \item{Parameter}{Parameter}
   \item{Unit}{Parameter unit}
   \item{Calculation}{How annual values should be calculated. Values: mean,
   min, max, median}
   \item{Min}{Minimum acceptable result}
   \item{Max}{Maximum acceptable result}
   \item{Excellent}{Excellent result}
   \item{Good}{Good result}
   \item{Fair}{Fair result}
   \item{Best}{Whether "Excellent" is lower or higher than "Fair". Values:
   high, low, NA}
 }
}
\usage{
dat_thresholds
}
\description{
Dataframe containing threshold values derived from the EPA goldbook and
state regulations.
}
\keyword{datasets}
