% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dropdown.R
\name{dropdown}
\alias{dropdown}
\title{Create dropdown menu}
\usage{
dropdown(
  id,
  label,
  choices,
  choice_names = NULL,
  sorted = TRUE,
  decreasing = FALSE,
  multiple = TRUE,
  max_options = NULL
)
}
\arguments{
\item{id}{String. Widget id.}

\item{label}{String. Widget heading/label.}

\item{choices}{List. Dropdown choices.}

\item{choice_names}{List. Display names for choices. Default `NULL`.}

\item{sorted}{Boolean. Whether to sort the choices. Default `TRUE`.}

\item{decreasing}{Boolean. Whether to sort choices in descending order.
Default `FALSE`.}

\item{multiple}{Boolean. Whether to allow multiple selections. Default
`TRUE`.}

\item{max_options}{Integer. Maximum number of selections. Default `NULL`.}
}
\value{
A dropdown widget.
}
\description{
`dropdown()` creates a dropdown widget.
}
