% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_embed.R
\name{embed_quarto}
\alias{embed_quarto}
\title{Embed quarto document}
\usage{
embed_quarto(path)
}
\arguments{
\item{path}{String. Path to exported Quarto .html document.}
}
\value{
Simple HTML
}
\description{
`embed_quarto()` converts a `Quarto` HTML document to simple
HTML that can be embedded in a Shiny app.
}
