% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_results.R
\name{format_results}
\alias{format_results}
\title{Format water quality data for wqdasbhoard}
\usage{
format_results(.data)
}
\arguments{
\item{.data}{Input dataframe}
}
\value{
Updated dataframe.
}
\description{
`format_results()` formats water quality data for use in
wqdashboard. Must run `qaqc_results()` first.
}
