% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sites.R
\name{format_sites}
\alias{format_sites}
\title{Format site metadata for use in wqdashboard}
\usage{
format_sites(.data)
}
\arguments{
\item{.data}{Input dataframe.}
}
\value{
Updated dataframe.
}
\description{
`format_sites()` formats site metadata for use in [wqdashboard].
}
\seealso{
[qaqc_sites()]
}
