% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_thresholds.R
\name{format_thresholds}
\alias{format_thresholds}
\title{Format threshold metadata for use in wqdashboard}
\usage{
format_thresholds(.data)
}
\arguments{
\item{.data}{Input dataframe}
}
\value{
Updated dataframe
}
\description{
`format_thresholds()` formats threshold metadata for use in
[wqdashboard].
}
