% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importwqd-package.R
\docType{package}
\name{importwqd-package}
\alias{importwqd}
\alias{importwqd-package}
\title{importwqd: Helper package for wqdashboard}
\description{
importwqd is a helper package for WQdashboard that contains functions and modules used to import and display water quality data.
}
\author{
\strong{Maintainer}: Mariel Sorlien \email{msorlien@nbep.org}

}
\keyword{internal}
