% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download.R
\name{mod_download_server}
\alias{mod_download_server}
\title{Download server}
\usage{
mod_download_server(id, sites, results, txt_citation, in_var)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_download_ui()`.}

\item{sites}{Dataframe. Site metadata.}

\item{results}{Dataframe. Result metadata.}

\item{txt_citation}{String. Text used for citation.}

\item{in_var}{Reactive output from `mod_sidebar_server`.}
}
\description{
`mod_download_server()` produces the server code for the
`wqdashboard` download tab.
}
