% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download.R
\name{mod_download_ui}
\alias{mod_download_ui}
\title{Download UI}
\usage{
mod_download_ui(id)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_download_server()`.}
}
\description{
`mod_download_ui()` produces the UI code for the `wqdashboard`
download tab.
}
