% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_compare.R
\name{mod_graph_compare_server}
\alias{mod_graph_compare_server}
\title{Comparison graph server}
\usage{
mod_graph_compare_server(id, df, group = "Site_Name", add_lines = FALSE)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_graph_compare_ui()`.}

\item{df}{Dataframe. Data to graph.}

\item{group}{String. Column used to group data.}

\item{add_lines}{Boolean. If `TRUE`, displays scatter plot as lines +
markers. If `FALSE`, displays scatter plot as markers only. Default `FALSE`.}
}
\description{
`mod_graph_compare_server()` produces the server code for a
graph and table comparing trends by group.
}
