% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_compare.R
\name{mod_graph_compare_ui}
\alias{mod_graph_compare_ui}
\title{Comparison graph UI}
\usage{
mod_graph_compare_ui(id)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_graph_compare_server()`.}
}
\description{
`mod_graph_compare_ui()` produces the UI code for a graph and
table comparing trends by group.
}
