% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph.R
\name{mod_graph_server}
\alias{mod_graph_server}
\title{Graph server}
\usage{
mod_graph_server(id, in_var)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_graph_ui()`.}

\item{in_var}{Reactive output from `mod_sidebar_server`.}
}
\description{
`mod_graph_server()` produces the server code for the `wqdashboard` graph
tab.
}
