% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_trend.R
\name{mod_graph_trend_server}
\alias{mod_graph_trend_server}
\title{Graph trends server}
\usage{
mod_graph_trend_server(id, df)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_graph_trend_ui()`.}

\item{df}{Dataframe. Data to graph.}
}
\description{
`mod_graph_trend_server()` produces the server code for a graph and
table showing long term trends for a single site, parameter, and depth.
}
