% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_trend.R
\name{mod_graph_trend_ui}
\alias{mod_graph_trend_ui}
\title{Graph trends UI}
\usage{
mod_graph_trend_ui(id)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_graph_trend_server()`.}
}
\description{
`mod_graph_trend_ui()` produces the UI code for a graph and
table showing long term trends for a single site, parameter, and depth.
}
