% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph.R
\name{mod_graph_ui}
\alias{mod_graph_ui}
\title{Graph UI}
\usage{
mod_graph_ui(id, varlist)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_graph_server()`.}

\item{varlist}{List of dropdown variables generated by `sidebar_var()`.}
}
\description{
`mod_graph_ui()` produces the UI code for the `wqdashboard`
graph tab.
}
