% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map.R
\name{mod_map_server}
\alias{mod_map_server}
\title{Map server}
\usage{
mod_map_server(
  id,
  sbVar,
  siteVar,
  df_raw,
  map_tab,
  shp_watershed = NULL,
  shp_river = NULL
)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_map_ui()`.}

\item{sbVar}{Reactive output from `mod_sidebar_server`.}

\item{siteVar}{Static list. Coordinate boundaries for site data.}

\item{df_raw}{Static dataframe. Default data.}

\item{map_tab}{Reactive boolean. If map tab is selected and data should be
updated, set to `TRUE`. Else, set to `FALSE`.}

\item{shp_watershed}{Static shapefile. Watershed polygons. Set to `NULL` if
no shapefile available. Default `NULL`.}

\item{shp_river}{Static shapefile. River polylines. Set to `NULL` if no
shapefile available. Default `NULL`}
}
\value{
Reactive list containing two items.
* graph_link: Button press for "View Trends"
* site: Site_ID for selected site when "View Trends" is pressed
}
\description{
`mod_map_server()` produces the map server for the
companion app`wqdashboard`.
}
