% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map.R
\name{mod_map_ui}
\alias{mod_map_ui}
\title{Map UI}
\usage{
mod_map_ui(id)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_map_server()`.}
}
\description{
`mod_map_ui()` produces the map UI for the companion app
`wqdashboard`.
}
