% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_server}
\alias{mod_report_server}
\title{Report card server}
\usage{
mod_report_server(id, in_var, df_raw)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_report_ui()`.}

\item{in_var}{Reactive output from `mod_sidebar_server`.}

\item{df_raw}{Dataframe. Default report card data.}
}
\description{
`mod_report_server()` produces the report card server for the
companion app `wqdashboard`.
}
