% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_ui}
\alias{mod_report_ui}
\title{Report card UI}
\usage{
mod_report_ui(id)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_report_server()`.}
}
\description{
`mod_report_ui()` produces the sidebar UI for the companion app
`wqdashboard`.
}
