% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_location.R
\name{mod_sidebar_location_ui}
\alias{mod_sidebar_location_ui}
\title{Sidebar location UI}
\usage{
mod_sidebar_location_ui(id, df_sites)
}
\arguments{
\item{id}{Namespace ID for module. Must match ID used by
`mod_sidebar_location_server()`.}

\item{df_sites}{Dataframe with site data. Must have been processed
by `format_sites()` and  include columns Site_ID, Site_Name.}
}
\description{
`mod_sidebar_location_ui()` is a helper module for
`mod_sidebar_ui()` that produces the UI used to filter and select sites.
}
\seealso{
mod_sidebar_location_server
}
