% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar.R
\name{mod_sidebar_ui}
\alias{mod_sidebar_ui}
\title{Sidebar UI}
\usage{
mod_sidebar_ui(id, df_data, df_score, df_sites)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_sidebar_server()`.}

\item{df_data}{Dataframe with result data. Must have been processed
by `format_results()` and include columns Parameter, Month, Year.}

\item{df_score}{Dataframe with data scores. Must have been processed
by `score_results()` and include column Parameter.}

\item{df_sites}{Dataframe with site data. Must have been processed
by `format_sites()` and  include columns Site_ID, Site_Name.}
}
\description{
`mod_sidebar_ui()` produces the sidebar UI for the companion app
`wqdashboard`.
}
