% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popup_text.R
\name{popup_text}
\alias{popup_text}
\title{Add column with popup text}
\usage{
popup_text(
  .data,
  col_list,
  col_title = NULL,
  target_col = "Popup",
  na_value = "-",
  hide_na = FALSE,
  style = "<b>in_title:</b> in_data"
)
}
\arguments{
\item{.data}{Dataframe.}

\item{col_list}{List or string. Columns used to populate values in popup
text. Will be added in order listed.}

\item{col_title}{List or string. The title to place at the start of each new
line. Default `NULL`.
* If list, must be same length as `col_list`.
* If `NULL`, `col_title` will be replaced with `col_list`. Underscores will
be replaced with spaces.}

\item{target_col}{String. Name of column to modify and add text to. If
column does not exist, it will be added to the dataframe. Default "Popup".}

\item{na_value}{String. Replacement for `NA` values. Default "-".}

\item{hide_na}{Boolean. If `TRUE`, will not add a new line of text if
`in_data` is `NA`. Default `TRUE`.}

\item{style}{String. How to format each line of data. String must include
"in_title" and "in_data" to represent the title and data value respectively.
Default "<b>in_title:</b> in_data".}
}
\value{
Updated dataframe with column named "Popup" unless `target_col`
set to different value. "Popup" column contains formatted popup text.
}
\description{
`popup_text()` generates and formats descriptive popup text for each
row in a dataframe. It updates `target_col` by concatenating the existing
text with the column name and value for each column in `col_list`.
* Unless `style` is updated, each column/value pair is formatted as
"<b>column name:</b> column value"
* A delimiter (`<br>`) is added between each line of text.
}
