% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_results.R
\name{preformat_results}
\alias{preformat_results}
\title{Pre-format Results}
\usage{
preformat_results(df, in_format)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
Updated dataframe.
}
\description{
Formats result data for WQdashboard by renaming columns,
  parameters, and qualifiers.
}
