% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_results.R
\name{prep_results}
\alias{prep_results}
\title{Prepare water quality data}
\usage{
prep_results(.data, df_colnames, df_param, df_unit, df_qual, df_activity)
}
\arguments{
\item{.data}{Input dataframe}

\item{df_colnames}{Dataframe. Must include paired list of column names in
columns "wqdashboard" and "Custom".}

\item{df_param}{Dataframe. Must include paired list of parameter names in
columns "wqdashboard" and "Custom".}

\item{df_unit}{Dataframe. Must include paired list of unit names in columns
"wqdashboard" and "Custom".}

\item{df_qual}{Dataframe. Must include paired list of qualifier names in
columns "wqdashboard" and "Custom".}

\item{df_activity}{Dataframe. Must include paired list of activity type names
in columns "wqdashboard" and "Custom".}
}
\value{
Updated dataframe
}
\description{
`prep_results()` prepares water quality data for use in
`wqdashboard` by updating column and variable names.
}
