% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sites.R
\name{prep_sites}
\alias{prep_sites}
\title{Prepare site metadata}
\usage{
prep_sites(.data, df_colnames)
}
\arguments{
\item{.data}{Input dataframe}

\item{df_colnames}{Dataframe. Must include paired list of column names in
columns "wqdashboard" and "Custom".}
}
\value{
Updated dataframe
}
\description{
`prep_sites()` prepares site data for use in `wqdashboard` by
updating colum names.
}
