% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_thresholds.R
\name{prep_thresholds}
\alias{prep_thresholds}
\title{Prepare threshold metadata}
\usage{
prep_thresholds(.data, df_param, df_unit)
}
\arguments{
\item{.data}{Input dataframe}

\item{df_param}{Dataframe. Must include paired list of parameter names in
columns "wqdashboard" and "Custom".}

\item{df_unit}{Dataframe. Must include paired list of unit names in columns
"wqdashboard" and "Custom".}
}
\value{
Updated dataframe
}
\description{
`prep_trhesholds()` prepares site data for use in `wqdashboard`
by updating parameter and unit names.
}
