% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_results.R
\name{qaqc_cat_results}
\alias{qaqc_cat_results}
\title{Check categorical water quality data for formatting errors}
\usage{
qaqc_cat_results(.data, sites)
}
\arguments{
\item{.data}{Input dataframe}

\item{sites}{Dataframe containing site metadata}
}
\value{
Updated dataframe
}
\description{
`qaqc_cat_results()` checks categorical water quality data for
major formatting errors. Specifically, it runs the following checks:
* Checks for missing values
* Checks for unknown sites
* Checks Depth column is numeric
* Checks Depth is in meters
* Assigns depth category
}
\seealso{
[qaqc_results()]
}
