% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_results.R
\name{qaqc_results}
\alias{qaqc_results}
\title{Check water quality data for formatting errors}
\usage{
qaqc_results(.data, site_data)
}
\arguments{
\item{.data}{Input dataframe}

\item{site_data}{Dataframe containing site metadata.}
}
\value{
Updated dataframe
}
\description{
`qaqc_results()` checks imported water quality data for major
formatting errors. Specifically, it runs the following checks:
* Checks for missing columns
* Checks for missing values
* Checks for unknown sites
}
\seealso{
[format_results()], [score_results()]
}
