% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_shapefile.R
\name{qaqc_shp}
\alias{qaqc_shp}
\title{Import shapefile}
\usage{
qaqc_shp(in_shp, field_name, path = NULL)
}
\arguments{
\item{in_shp}{String. Name of shapefile.}

\item{field_name}{String. Name of field that contains polygon names. Text in
this field will be used for hovertext and alt text.}

\item{path}{String. Path to shapefile.}
}
\value{
Updated shapefile
}
\description{
`qaqc_shp()` prepares shapefiles for `wqdashboard`. Shapefiles
must use the `WGS 84` projection.
}
