% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sites.R
\name{qaqc_sites}
\alias{qaqc_sites}
\title{Check site metadata for formatting errors}
\usage{
qaqc_sites(.data)
}
\arguments{
\item{.data}{Input dataframe}
}
\value{
Updated dataframe
}
\description{
`qaqc_sites()` checks imported site metadata for major
formatting errors. Specifically, it runs the following checks:
* Checks for missing columns
* Checks for missing values
* Checks for duplicate site ID, site name, coordinates
* Checks if coordinate, depth columns contain non-numeric values
* Checks if depth values are illogical (eg surface depth is greater than
midwater depth)
}
\seealso{
[format_sites()]
}
