% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_thresholds.R
\name{qaqc_thresholds}
\alias{qaqc_thresholds}
\title{Check thresholds metadata for formatting errors}
\usage{
qaqc_thresholds(.data)
}
\arguments{
\item{.data}{Input dataframe}
}
\value{
Updated dataframe
}
\description{
`qaqc_thresholds()` checks imported threshold metadata for major
formatting errors. Specifically, it runs the following checks:
* Checks for missing columns
* Checks for missing values
* Checks if rows contain logical threshold values
}
