% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_results.R
\name{score_results}
\alias{score_results}
\title{Calculate annual scores and format results for wqdashboard}
\usage{
score_results(.data, sites)
}
\arguments{
\item{.data}{Input dataframe}

\item{sites}{Dataframe containing site metadata}
}
\value{
Updated dataframe
}
\description{
`score_results()` calculates annual scores and formats data
for use in [wqdashboard]. Must run `format_results()` first.
}
