% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_results.R
\name{sidebar_var}
\alias{sidebar_var}
\title{Generate dropdown lists for sidebar}
\usage{
sidebar_var(df_sites, df_data, df_score)
}
\arguments{
\item{df_sites}{Dataframe with site data. Must have been processed
by `format_sites()` and  include columns Site_ID, Site_Name.}

\item{df_data}{Dataframe with result data. Must have been processed
by `format_results()` and include columns Parameter, Month, Year.}

\item{df_score}{Dataframe with data scores. Must have been processed
by `score_results()` and include column Parameter.}
}
\value{
Named list containing the following:
* state - List of state names.
* town - List of town names.
* watershed - List of watershed names.
* param - List of parameters
* param_short - List of parameters that include at least one categorical
score (eg excellent, good, fair)
* depth - List of depths
* year - Sorted list of years
* month - Continuous, sorted list of months
}
\description{
`sidebar_var()` processes site and result data to generate
a list of dropdown options for `mod_sidebar_ui()`.
}
