% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_units.R
\name{standardize_detection_units}
\alias{standardize_detection_units}
\title{Standardize detection limit units}
\usage{
standardize_detection_units(.data)
}
\arguments{
\item{.data}{Dataframe}
}
\value{
Updated dataframe. Detection limit and detection limit unit will
be updated to match that row's result unit.
}
\description{
`standardize_detection_units()` updates result data so that
each row uses the same units for results and detection limits.
}
