% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_units.R
\name{standardize_result_units}
\alias{standardize_result_units}
\title{Standardize result units}
\usage{
standardize_result_units(.data)
}
\arguments{
\item{.data}{Dataframe}
}
\value{
Updated dataframe. Unit and result values will be updated to ensure
each parameter uses a single, consitent unit.
}
\description{
`standardize_result_units()` updates result data so that each
parameter uses a single, standard unit for all results.
}
