% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_units.R
\name{standardize_threshold_units}
\alias{standardize_threshold_units}
\title{Standardize threshold units}
\usage{
standardize_threshold_units(.data, result_data)
}
\arguments{
\item{.data}{Dataframe}

\item{result_data}{Dataframe containing result data.}
}
\value{
Updated dataframe. Unit and threshold values will be updated to
match the units used in `result_data`.
}
\description{
`standardize_threshold_units()` updates threshold data so that
each parameter uses the same units as those used in `result_data`.
}
