% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqdutils-package.R
\docType{package}
\name{wqdutils-package}
\alias{wqdutils}
\alias{wqdutils-package}
\title{wqdutils: Helper package for wqdashboard}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Mariel Sorlien \email{msorlien@nbep.org}

}
\keyword{internal}
