% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap Text}
\usage{
wrap_text(old_list, str_len = 20, linebreak = NULL)
}
\arguments{
\item{old_list}{Input list.}

\item{str_len}{Length of string.}

\item{linebreak}{Linebreak symbol.}
}
\value{
Updated list.
}
\description{
Wrap long strings of text.
}
