% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculate annual parameter score}
\usage{
calculate_score(
  site_id,
  parameter,
  unit,
  depth = NA,
  score_max,
  score_min,
  score_mean,
  score_median
)
}
\arguments{
\item{site_id}{Site ID.}

\item{parameter}{Parameter.}

\item{unit}{Parameter unit.}

\item{depth}{Depth category. Default value NA.}

\item{score_max}{Maximum score.}

\item{score_min}{Minimum score.}

\item{score_mean}{Average score.}

\item{score_median}{Median score.}
}
\value{
List including a numeric score (score_num), way the numeric score was
calculated (score_typ), and a categorical score (score_str).
}
\description{
`calculate_score()` calculates annual numeric and categorical
scores for each parameter/site/depth combination.

The numeric score returns the annual minimum, maximum, median, or mean value.
If the threshold table does not specify which value to return, the mean value
is used.

The categorical score compares the numeric score to the threshold table and
returns a value of Excellent, Good, Fair, Poor, Does Not Meet Criteria, or
Meets Criteria.
}
\seealso{
score_results()
}
