% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_results.R
\name{format_results}
\alias{format_results}
\title{Format Results}
\usage{
format_results(df)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
Updated dataframe.
}
\description{
Formats water quality data for use in app. Must run
  `QAQC_results` first.
}
