% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sites.R
\name{format_sites}
\alias{format_sites}
\title{Format site metadata for use in wqdashboard}
\usage{
format_sites(.data, site_list = NULL)
}
\arguments{
\item{.data}{Input dataframe}

\item{site_list}{List of site ID values.  Dataframe will be filtered for
matching values in column "Site_ID". If `NULL`, dataframe will not be
filtered. Default `NULL`.}
}
\value{
Updated dataframe.
}
\description{
`format_sites()` formats site metadata for use in [wqdashboard].
}
\seealso{
[qaqc_sites()]
}
