% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_trend.R
\name{mod_graph_trend_server}
\alias{mod_graph_trend_server}
\title{Graph trends server}
\usage{
mod_graph_trend_server(
  id,
  df,
  in_var = list(lines = TRUE, trend = TRUE, tresh = TRUE)
)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_graph_trend_ui()`.}

\item{df}{Dataframe. Data to graph.}

\item{in_var}{List with three items: style, trend, thresh
* `lines`: Boolean. If `TRUE`, displays scatter plot as lines +
markers. If `FALSE`, displays scatter plot as markers only.
* `trend`: Boolean. If `TRUE`, adds trendline.
* `thresh`: Boolean. If `TRUE`, adds red bar to indicate values outside
acceptable range and a blue bar to indicate excellent values. Default
`FALSE`. description}
}
\description{
`mod_graph_trend_server()` produces the server code for a graph and
table showing long term trends for a single site, parameter, and depth.
}
