% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map.R
\name{mod_map_server}
\alias{mod_map_server}
\title{Map server}
\usage{
mod_map_server(
  id,
  in_var,
  map_bounds,
  df_raw,
  selected_tab,
  shp_watershed = NULL,
  shp_river = NULL
)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_map_ui()`.}

\item{in_var}{Reactive output from `mod_sidebar_server`.}

\item{map_bounds}{Numeric list. Map boundaries. Should be ordered minimum
longitude, minimum latitude, maximum longitude, maximum latitude.}

\item{df_raw}{Dataframe. Default map data.}

\item{selected_tab}{Reactive string. Name of selected tab.}

\item{shp_watershed}{Shapefile. Watershed polygons. Set to `NULL` if
no shapefile available. Default `NULL`.}

\item{shp_river}{Shapefile. River polylines. Set to `NULL` if no
shapefile available. Default `NULL`.}
}
\value{
Reactive list containing two items.
* graph_link: Button press for "View Trends"
* site: Site_ID for selected site when "View Trends" is pressed
}
\description{
`mod_map_server()` produces the map server for the
companion app`wqdashboard`.
}
