% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_location.R
\name{mod_sidebar_location_server}
\alias{mod_sidebar_location_server}
\title{Sidebar location Server}
\usage{
mod_sidebar_location_server(id, sites, tab, selected_site)
}
\arguments{
\item{id}{Internal parameter for [shiny].}

\item{sites}{Dataframe with site data. Must include columns Site_ID,
Site_Name. Optionally may include columns State, Town, Watershed.
Default `NULL`.}

\item{tab}{Reactive variable. Name of selected tab.}

\item{selected_site}{Reactive variable. Site_ID for site selected in
`mod_map_ui()`.}
}
\description{
A shiny Module used to filter data by location. Submodule for
`mod_sidebar_ui()`.
}
\seealso{
mod_sidebar_location_ui
}
