% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_location.R
\name{mod_sidebar_location_ui}
\alias{mod_sidebar_location_ui}
\title{Sidebar location UI}
\usage{
mod_sidebar_location_ui(id, sites)
}
\arguments{
\item{id}{Internal parameter for [shiny].}

\item{sites}{Dataframe with site data. Must include columns Site_ID,
Site_Name. Optionally may include columns State, Town, Watershed.
Default `NULL`.}
}
\description{
A shiny Module used to filter data by location. Submodule for
`mod_sidebar_ui()`.
}
\seealso{
mod_sidebar_location_server
}
