% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar.R
\name{mod_sidebar_server}
\alias{mod_sidebar_server}
\title{Sidebar server}
\usage{
mod_sidebar_server(
  id,
  df_sites,
  df_data,
  df_score,
  selected_tab,
  selected_site
)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_sidebar_ui()`.}

\item{df_sites}{Dataframe with site data. Must have been processed
by `format_sites()` and  include columns Site_ID, Site_Name.}

\item{df_data}{Dataframe with result data. Must have been processed
by `format_results()` and include columns Parameter, Month, Year.}

\item{df_score}{Dataframe with data scores. Must have been processed
by `score_results()` and include column Parameter.}

\item{selected_tab}{Reactive variable. Name of selected tab.}

\item{selected_site}{Reactive variable. Site_ID for site selected in
`mod_map_ui()`.}
}
\value{
TO DO.
}
\description{
`mod_sidebar_server()` produces the sidebar server for the
companion app`wqdashboard`.
}
