% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar.R
\name{mod_sidebar_ui}
\alias{mod_sidebar_ui}
\title{Sidebar UI}
\usage{
mod_sidebar_ui(id, varlist)
}
\arguments{
\item{id}{Namespace ID for module. Should match ID used by
`mod_sidebar_server()`.}

\item{varlist}{List of dropdown variables generated by `sidebar_var()`.}
}
\description{
`mod_sidebar_ui()` produces the sidebar UI for the companion app
`wqdashboard`.
}
