% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_results.R
\name{qaqc_results}
\alias{qaqc_results}
\title{QAQC Results}
\usage{
qaqc_results(df, date_format = NULL)
}
\arguments{
\item{df}{Input dataframe.}

\item{date_format}{Date format as string.}
}
\value{
Updated dataframe.
}
\description{
Run quick quality control check on water quality data. Runs the
  following checks:
  * Checks all mandatory columns present
  * Checks for missing values
  * Formats dates
  * Formats units
}
