% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_results.R
\name{qaqc_results}
\alias{qaqc_results}
\title{Check water quality data for formatting errors}
\usage{
qaqc_results(.data, sites)
}
\arguments{
\item{.data}{Input dataframe}

\item{sites}{Dataframe containing site metadata}
}
\value{
Updated dataframe
}
\description{
`qaqc_results()` checks imported water quality data for major
formatting errors. Specifically, it runs the following checks:
* Checks for missing values
* Checks for unknown sites
* Checks Results, Depth columns are numeric
* Checks Depth is in meters
* Assigns depth category
* Standardizes units
}
\seealso{
[format_results()], [score_results()]
}
