% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_results.R
\name{score_results}
\alias{score_results}
\title{Format df_score}
\usage{
score_results(df)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
Updated dataframe.
}
\description{
Formats water quality data for use in app. Must run
  `format_results` first.
}
